use floats instead of double

+++ libc/baselibc/src/tinyprintf.c
@@ -242,7 +242,7 @@ size_t tfp_format(FILE *putp, const char *fmt, va_list va)
     char lng;
     void *v;
 #if MYNEWT_VAL(FLOAT_USER)
-    double d;
+    float d;
     int n;
 #endif
     int i;
@@ -356,7 +356,7 @@ size_t tfp_format(FILE *putp, const char *fmt, va_list va)
 #if MYNEWT_VAL(FLOAT_USER)
             case 'f':
                 p.base = 10;
-                d = va_arg(va, double);
+                d = (float)va_arg(va, double);
                 /* Convert to an int to get the integer part of the number. */
                 n = d;
                 /* Convert to ascii */
@@ -365,7 +365,7 @@ size_t tfp_format(FILE *putp, const char *fmt, va_list va)
                  * towards 0.  If the number is in the range (-1, 0), the
                  * negative sign was lost.  Preserve the sign in this case.
                  */
-                if (d < 0.0) {
+                if ((float)d < 0.0f) {
                     p.sign = 1;
                 }
                 /* Ignore left align for integer part */
@@ -379,7 +379,7 @@ size_t tfp_format(FILE *putp, const char *fmt, va_list va)
                 /* Write integer part to console */
                 written += putchw(putp, &p);
                 /* Take the decimal part and multiply by 1000 */
-                n = (d-n)*1000;
+                n = (d-n)*1000.0f;
                 /* Convert to ascii */
                 i2a(n, &p);
                 /* Set the leading zeros for the next integer output to 3 */
